Move preferences to its own activity, leaving a "blank" openkeychain api one
