+++ app/src/main/AndroidManifest.xml
@@ -61,6 +61,13 @@
 					   android:value="com.zeapo.pwdstore.PasswordStore" />
 		</activity>
 
+		<activity android:name=".autofill.AutofillPreferenceActivity">
+			android:parentActivityName=".PasswordStore">
+
+			<meta-data android:name="android.support.PARENT_ACTIVITY"
+					   android:value="com.zeapo.pwdstore.PasswordStore" />
+		</activity>
+
         <activity android:name="net.rdrei.android.dirchooser.DirectoryChooserActivity" />
 	</application>
